perf: use the faster grayscale converter
This promotes the "fast" version to the only version and removes the benchmark test.
